Add files via upload
This zipped folder contains the main program, classes, images, and sounds for the game. It should be run through Visual Studio and MonoGame. After extracting the file, the easiest way to run the game is by going to: File --> Open... --> Open Folder, then selecting the folder. The main file to run once the folder is open in Visual Studio is Game1.cs